general class for tasks
